QLocale/Win: statically assert that is_sorted(windows_to_iso_list)
Consequently, remove the corresponding comment.

Need to mark the array as constexpr in order to run q20::is_sorted on
it.

Instead of overloading op< (which would be wrong, because there's no
one natural ordering of these entries), write the idiomatic manual
mixed-mode comparator By<FieldName> and use that. A follow-up commit
will use the same function object to replace the hand-written binary
search with a call to lower_bound().

Task-number: QTBUG-103721
Change-Id: I47743b5620dc3cece73b2e5bae658d5a636554dd
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>