QProcess: Extend the setStandardOutputFile doc
It seems to be a hidden gem that one can stop redirecting the standard
output to a file without creating a new QProcess instance.

I have discovered this while browsing the private implementation of
QProcess trying to answer this question:

https://stackoverflow.com/a/72989131/2682142

I believe that this should be documented unless it is not meant to be
guaranteed behavior.

Change-Id: I180764d4cc56f6395d6d65942077e3ede63c71ea
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>