Fix two trivial defects in testlib docs
A fragment of inline code wasn't wrapped in \c{}.
Doing so made the quotes it was wrapped in redundant.

Documentation of of QSKIP's parameter claimed it does something it
should, but it's up to the client code using it to write a description
that lives up to that.

Change-Id: Ib355b9b85c9986f244ae01479134245f182c912c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 6dd6bf3467040b3efceafa317de6e833ee5ec15e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>