QVariant: update the noexcept content for a few types
Commit 2f0a625fd4036b71286dfa7de4c9d795025164e7 added noexcept for
these, but didn't verify that the operation itself was noexcept. And it
wasn't on 32-bit systems, because sizeof(void *) is only 4 bytes, making
QVariant and QVariant::Private a mere 12 bytes. That's insufficient for
QUuid and for almost all geometric types when qreal==double.

We can't use sizeof() in qvariant.h because most of those classes are
only forward-declared.

Change-Id: I6f936da6f6e84d649f70fffd1705ce948891d06a
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>