Update QTimeZone benchmark's list of tested zones
It purports to be a list of those named in QDateTime's tests, but was
a bit out of date. In the process, sort them into somewhat coherent
order (to make it easier to verify whether one is missing next time I
check) and reformat.

Change-Id: I16e7ded6f8b00e226513bd06d6174a79f7a0c675
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 0ed89870788f28daf1504a6204343d6b158b2cf3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>