Sync Vulkan device features between QVulkanWindow and QRhi
Enable all 1.0 core features that are reported as supported.
This has been recently added for QVulkanWindow, but not for QRhi's
Vulkan backend. Synchronize the behavior between the two.

In addition, explicitly turn off robustBufferAccess in both places.
This is a feature that may affect performance by nature, and so we
cannot just blindly enable it.

Amends 7888d846ebd337d586971aa92bbd21cd227c2bc5

Task-number: QTBUG-105158
Change-Id: Icd7d1800a24a39a2bc427966022185885fa92d55
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>